agregando funcionalidades a post.delete, mas controles de autorizacion

+++ server/config/db_post.py
@@ -19,10 +19,15 @@ def post_post(db: Session, request: PostBase):
 def get_all_post(db: Session):
     return db.query(DbPost).all()
 
-def delete_post(db: Session, id: int):
+def get_post_byt_id(db: Session, id:int):
+    return db.query(DbPost).filter(DbPost.id == id).first()
+
+def delete_post(db: Session, id: int, user_id: int):
     post = db.query(DbPost).filter(DbPost.id == id).first()
     if not post:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'No pudimos encontrar el post con id:    {id}, intentelo nuevamente. ')
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f'No pudimos encontrar el post con id: {id}, intentelo nuevamente. ')
+    if post.user_id != user_id:
+        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail='Solo puedes borrar post de los que seas propietario.')
     db.delete(post)
     db.commit()
     return 'El post fue eliminado correctamente.'
